Move Header and remove SerachFilter (duplicated in ol-search-ui)
